// app/lib/types.ts
export interface Page {
// ...
  faqSection?: {
    enabled: boolean;
    title?: any;
    description?: any;
    items: Array<{
      question?: any;
      answer?: any;
    }>;
  };
// ...
}